+++ src/index.ts
@@ -37,6 +37,7 @@ export * from './new-components/avatar'
 export * from './new-components/badge'
 export * from './new-components/modal'
 export * from './new-components/tabs'
+export * from './new-components/tooltip'
 
 // hooks
 export * from './hooks/use-previous'
@@ -50,8 +51,6 @@ export { default as KeyboardShortcut } from './components/keyboard-shortcut'
 export { default as KeyCapturer, SUPPORTED_KEYS } from './components/key-capturer'
 export { default as ProgressBar } from './components/progress-bar'
 export { default as Time } from './components/time'
-export { Tooltip } from './components/tooltip'
-export type { TooltipProps } from './components/tooltip'
 export * from './components/menu'
 
 //