+++ src/new-components/password-field/password-field.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { useForkRef } from 'ariakit-utils'
-import { Tooltip } from '../../components/tooltip'
+import { Tooltip } from '../tooltip'
 import { BaseField, BaseFieldVariantProps } from '../base-field'
 import { Box } from '../box'
 import { useId } from '../common-helpers'