+++ src/new-components/tooltip/tooltip.tsx
@@ -1,23 +1,79 @@
 import * as React from 'react'
-import classNames from 'classnames'
 
 import {
     useTooltipState as useAriakitTooltipState,
-    TooltipStateProps as AriakitTooltipStateProps,
     Tooltip as AriakitTooltip,
-    TooltipProps as AriakitTooltipProps,
     TooltipAnchor,
+    TooltipArrow,
+} from 'ariakit/tooltip'
+import { Box } from '../box'
+
+import type {
+    TooltipStateProps as AriakitTooltipStateProps,
     TooltipAnchorProps,
 } from 'ariakit/tooltip'
 import type { PopoverState } from 'ariakit/popover'
 
-import './tooltip.less'
+import styles from './tooltip.module.css'
 
-type TooltipProps = Omit<AriakitTooltipProps, 'children' | 'state'> & {
+type TooltipProps = {
+    /**
+     * The element that triggers the tooltip. Generally a button or link.
+     *
+     * It should be an interactive element accessible both via mouse and keyboard interactions.
+     */
     children: React.ReactNode
+
+    /**
+     * The content to show in the tooltip.
+     *
+     * It can be rich content provided via React elements, or string content. It should not include
+     * interactive elements inside it. This includes links or buttons.
+     *
+     * You can provide a function instead of the content itself. In this case, the function should
+     * return the desired content. This is useful if the content is expensive to generate. It can
+     * also be useful if the content dynamically changes often, so every time you trigger the
+     * tooltip the content may have changed (e.g. if you show a ticking time clock in the tooltip).
+     *
+     * The trigger element will be associated to this content via `aria-describedby`. This means
+     * that the tooltip content will be read by assistive technologies such as screen readers. It
+     * will likely read this content right after reading the trigger element label.
+     */
     content: React.ReactNode | (() => React.ReactNode)
+
+    /**
+     * How to place the tooltip relative to its trigger element.
+     *
+     * The possible values are "top", "bottom", "left", "right". Additionally, any of these values
+     * can be combined with `-start` or `-end` for even more control. For instance `top-start` will
+     * place the tooltip at the top, but with the start (e.g. left) side of the toolip and the
+     * trigger aligned. If neither `-start` or `-end` are provided, the tooltip is centered along
+     * the vertical or horizontal axis with the trigger.
+     *
+     * The position is enforced whenever possible, but tooltips can appear in different positions
+     * if the specified one would make the tooltip intersect with the viewport edges.
+     *
+     * @default 'top'
+     */
     position?: PopoverState['placement']
+
+    /**
+     * The separation (in pixels) between the trigger element and the tooltip.
+     * @default 3
+     */
     gapSize?: number
+
+    /**
+     * Whether to show an arrow-like element attached to the tooltip, and pointing towards the
+     * trigger element.
+     * @default false
+     */
+    withArrow?: boolean
+
+    /**
+     * An escape hatch, in case you need to provide a custom class name to the tooltip.
+     */
+    exceptionallySetClassName?: string
 }
 
 // These are exported to be used in the tests, they are not meant to be exported publicly
@@ -42,8 +98,8 @@ function Tooltip({
     content,
     position = 'top',
     gapSize = 3,
-    className,
-    ...props
+    withArrow = false,
+    exceptionallySetClassName,
 }: TooltipProps) {
     const state = useDelayedTooltipState({ placement: position, gutter: gapSize })
 
@@ -63,28 +119,27 @@ function Tooltip({
      * Prevents the tooltip from automatically firing on focus all the time. This is to prevent
      * tooltips from showing when the trigger element is focused back after a popover or dialog that
      * it opened was closed. See link below for more details.
-     * @see https://github.com/reakit/reakit/discussions/749
+     * @see https://github.com/ariakit/ariakit/discussions/749
      */
     function handleFocus(event: React.FocusEvent<HTMLDivElement>) {
-        // If focus is not followed by a key up event, does it mean that it's not
-        // an intentional keyboard focus? Not sure but it seems to work.
-        // This may be resolved soon in an upcoming version of reakit:
-        // https://github.com/reakit/reakit/issues/750
-        function handleKeyUp(e: Event) {
-            const eventKey = (e as KeyboardEvent).key
+        // If focus is not followed by a key up event, does it mean that it's not an intentional
+        // keyboard focus? Not sure but it seems to work.
+        // This may be resolved soon in an upcoming version of ariakit:
+        // https://github.com/ariakit/ariakit/issues/750
+        function handleKeyUp(event: Event) {
+            const eventKey = (event as KeyboardEvent).key
             if (eventKey !== 'Escape' && eventKey !== 'Enter' && eventKey !== 'Space') {
                 state.show()
             }
         }
         event.currentTarget.addEventListener('keyup', handleKeyUp, { once: true })
-        // Prevent tooltip.show from being called by TooltipReference
-        event.preventDefault()
+        event.preventDefault() // Prevent tooltip.show from being called by TooltipReference
         child?.props?.onFocus?.(event)
     }
 
     return (
         <>
-            <TooltipAnchor state={state} ref={child.ref} onFocus={handleFocus}>
+            <TooltipAnchor state={state} ref={child.ref} onFocus={handleFocus} described>
                 {(anchorProps: TooltipAnchorProps) => {
                     // Let child props override anchor props so user can specify attributes like tabIndex
                     // Also, do not apply the child's props to TooltipAnchor as props like `as` can create problems
@@ -93,13 +148,22 @@ function Tooltip({
                 }}
             </TooltipAnchor>
             {state.visible ? (
-                <AriakitTooltip
-                    {...props}
+                <Box
+                    as={AriakitTooltip}
                     state={state}
-                    className={classNames('reactist_tooltip', className)}
+                    className={[styles.tooltip, exceptionallySetClassName]}
+                    background="toast"
+                    borderRadius="standard"
+                    paddingX="small"
+                    paddingY="xsmall"
+                    maxWidth="medium"
+                    width="fitContent"
+                    overflow="hidden"
+                    textAlign="center"
                 >
+                    {withArrow ? <TooltipArrow /> : null}
                     {typeof content === 'function' ? content() : content}
-                </AriakitTooltip>
+                </Box>
             ) : null}
         </>
     )